affafasfd

+++ api/authentication/models.py
@@ -2,24 +2,41 @@ from django.db import models
 from django.utils.translation import gettext_lazy as _
 from django.contrib.auth.models import AbstractBaseUser, PermissionsMixin
 from django.utils import timezone
-from random import randint
-import os
 from django.conf import settings
+from django.utils.deconstruct import deconstructible
+import os
 
 from .manager import UserManager
 
 
-def upload_avatar_to(prefix):
-    def _path(instance, filename):
-        format = filename.split(".")[-1]
-        path = os.path.join(settings.MEDIA_ROOT, prefix, instance.email)
+# @deconstructible
+# class UploadAvatarTo:
+#     def __init__(self, sub_path):
+#         self.path = sub_path
 
+#     def __call__(self, instance, filename):
+#         format = filename.split(".")[-1]
+#         path = os.path.join(settings.MEDIA_ROOT, "users/", instance.email)
+
+#         print(path)
+
+#         for file in os.listdir(path):
+#             if "avatar" in file:
+#                 os.remove(os.path.join(path, file))
+
+#         return os.path.join(path, f"avatar.{format}")
+
+
+def _path(instance, filename):
+    format = filename.split(".")[-1]
+    path = os.path.join(settings.MEDIA_ROOT, "users/", instance.email)
+
+    if os.path.exists(path):
         for file in os.listdir(path):
             if "avatar" in file:
                 os.remove(os.path.join(path, file))
 
-        return os.path.join(path, f"avatar.{format}")
-    return _path
+    return os.path.join(path, f"avatar.{format}")
 
 
 ROLES = (('a', 'admin'), ('m', 'manager'), ('d', 'driver'), ('c', 'client'))
@@ -51,8 +68,11 @@ class User(AbstractBaseUser, PermissionsMixin):
         _('Подтвержден'), default=False,
     )
     avatar = models.ImageField(
-        _('Аватар'), upload_to=upload_avatar_to("users/"),
+        _('Аватар'), upload_to=_path,
         null=True, blank=True
+    )
+    is_online = models.BooleanField(
+        _('Онлайн'), default=False
     )
     
     is_staff = models.BooleanField(_("Сотрудник"), default=False)
